moved fixtures to conftest.py, changed entities/run_test.py so it uses fixtures, deleted unused imports Signed-off-by: Wiktor Wasowicz wiktor7645@gmail.com
